Split store byte range helpers
